Wave 18: code review remediation (S63-S67)
S63 — Security Hardening: const-time comparisons (subtle crate), auth rate limiting (5/min login, 1/hr register), token TTL enforcement, unified error messages (no enumeration), /content/publish auth, dev-login gated behind REACHLOCK_DEV_MODE. 14 findings fixed.

S64 — Server Robustness: player_senders race (session nonce tagging), PresenceManager broadcast deadlock fix (clone before await), LLM success metric fix, quota TOCTOU, PgSeedStore panic on empty SELECT, heartbeat on any message, atomic snapshots, WS close frame on auth fail, query-string percent-decoding, MemoryPlayerStore lock ordering, TURN HMAC-SHA256 upgrade. 15 findings fixed.

S65 — Client Critical Fixes: duplicate lighting removed, Hull reset on respawn (invincibility fix), SfxQueue double-init, audio entity cleanup system, VoiceManager Drop timeout, mic stream proper lifecycle, synthesis rate limit, contract library sort dedup, reputation keybind conflict, settings_ui O(n) fix, DOCK_RADIUS FIXME, minimum respawn fee. 14 findings fixed.

S66 — Editor & CLI Quality: save_all Ok(false)/Ok(true)/Err distinction, HTTP error body fix, duplicate GoodId detection, active_tab documentation, hull module cleanup, blank_file random seed. 6 findings fixed.

S67 — Core Determinism & Quality: f64→int in theater, SystemTime parameterized in log/metadata, HashMap→BTreeMap in 4 structs, FactionMap→BTreeMap, f32→Fixed in faction+sfx, SeededRng Clone/Copy, Vec→BTreeSet for fired_chapters, Serialize/Deserialize on GeneratedAudio, deny_unknown_fields on contracts, include_str path fix, gallery/universe casts. 16 findings fixed.

Total: 65 findings addressed. 441 core + 40 server + 14 editor tests pass. All crates clippy clean.

+++ reachlock-core/src/galaxy/coord.rs
@@ -9,7 +9,7 @@ use crate::universe::tier::UniverseTier;
 /// the loop across the two public functions.
 fn coord_fnv1a(coord: &GalaxyCoord, universe: UniverseTier) -> u64 {
     let mut h = 0xCBF2_9CE4_8422_2325u64;
-    for v in [coord.x, coord.y, coord.z, universe as u8 as i64] {
+    for v in [coord.x, coord.y, coord.z, universe as i64] {
         for b in v.to_le_bytes() {
             h ^= b as u64;
             h = h.wrapping_mul(0x0000_0100_0000_01B3);
